Refactor combineTwoList into ListUtil utility file
